+++ cstar/tests/integration_tests/conftest.py
@@ -1,6 +1,6 @@
-from cstar.tests.roms.fixtures import fetch_roms_tools_source_data  # noqa: F401
-from cstar.tests.roms.fixtures import fetch_remote_test_case_data  # noqa: F401
-from cstar.tests.blueprints.fixtures import modify_template_blueprint  # noqa : F401
+from cstar.tests.integration_tests.fixtures import fetch_roms_tools_source_data  # noqa: F401
+from cstar.tests.integration_tests.fixtures import fetch_remote_test_case_data  # noqa: F401
+from cstar.tests.integration_tests.blueprints.fixtures import modify_template_blueprint  # noqa : F401
 
 import builtins
 from contextlib import contextmanager
